hw_4_kafka/producer/requirements.txt
fastapi[standard]
uvicorn[standard]
kafka-python